Extract Home graphql types to graphql.ts
