+++ src/Home/Home.tsx
@@ -1,38 +1,12 @@
 import React from 'react'
-import gql from 'graphql-tag'
 import { useQuery } from '@apollo/react-hooks'
 
+import { USER_QUERY, UserQuery } from './graphql'
 import RoomSelector from './RoomSelector'
 import TeamSelector from './TeamSelector'
 
-import { UserType } from 'lib/apiTypes'
-
-type UserQuery = {
-  user: UserType
-}
-const USER_QUERY = gql`
-  query UserQuery {
-    user {
-      id
-      email
-      name
-      activeRoom {
-        id
-      }
-      activeTeam {
-        id
-        name
-      }
-      teams {
-        id
-        name
-      }
-    }
-  }
-`
-
 const Home: React.FC = () => {
-  const { loading, error, data } = useQuery<UserQuery>(USER_QUERY)
+  const { loading, error, data } = useQuery<UserQuery['data']>(USER_QUERY)
 
   if (loading) {
     return <p>Loading</p>
